feat: enable users to skip weeks

+++ src/main/java/es/sralloza/choremanagementbot/controllers/TenantsController.java
@@ -1,6 +1,8 @@
 package es.sralloza.choremanagementbot.controllers;
 
 import es.sralloza.choremanagementbot.models.db.DBTenant;
+import es.sralloza.choremanagementbot.models.io.TenantCreate;
+import es.sralloza.choremanagementbot.repositories.db.DBSkippedWeekRepository;
 import es.sralloza.choremanagementbot.repositories.db.DBTenantsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,6 +15,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 import static org.springframework.http.HttpStatus.NO_CONTENT;
 
@@ -21,6 +25,8 @@ import static org.springframework.http.HttpStatus.NO_CONTENT;
 public class TenantsController {
     @Autowired
     private DBTenantsRepository dbTenantsRepository;
+    @Autowired
+    private DBSkippedWeekRepository dbSkippedWeekRepository;
 
     @GetMapping()
     public List<DBTenant> listTenants() {
@@ -28,7 +34,8 @@ public class TenantsController {
     }
 
     @PostMapping()
-    public DBTenant createTenant(@RequestBody DBTenant tenant) {
+    public DBTenant createTenant(@RequestBody TenantCreate tenantCreate) {
+        var tenant = new DBTenant(tenantCreate.getTelegramId(), tenantCreate.getUsername(), UUID.randomUUID());
         return dbTenantsRepository.save(tenant);
     }
 
@@ -36,5 +43,9 @@ public class TenantsController {
     @ResponseStatus(value = NO_CONTENT)
     public void deleteTenant(@PathVariable("id") Integer tenantId) {
         dbTenantsRepository.deleteById(tenantId);
+        var skippedWeeks = dbSkippedWeekRepository.findAll().stream()
+                .filter(week -> week.getTenantId().equals(tenantId))
+                .collect(Collectors.toList());
+        dbSkippedWeekRepository.deleteAll(skippedWeeks);
     }
 }
